split up impl and test
